# Use Radix Trie to replace BTreeSet