Before:

```
running 7 tests
test create_ip_range_with_chnlists ... bench:   8,228,000 ns/iter (+/- 274,270)
test new_ip_range                  ... bench:         608 ns/iter (+/- 10)
test parse_chnlists                ... bench:     741,950 ns/iter (+/- 222,449)
test parse_one_ip                  ... bench:          21 ns/iter (+/- 0)
test parse_one_subnet1             ... bench:         106 ns/iter (+/- 38)
test parse_one_subnet2             ... bench:          96 ns/iter (+/- 7)
test test_10000_ips_in_chnlists    ... bench:   6,923,600 ns/iter (+/- 1,828,400)
```

After:

```
test create_ip_range_with_chnlists ... bench:   2,851,100 ns/iter (+/- 3,460,050)
test new_ip_range                  ... bench:           1 ns/iter (+/- 0)
test parse_chnlists                ... bench:     743,600 ns/iter (+/- 13,600)
test parse_one_ip                  ... bench:          21 ns/iter (+/- 0)
test parse_one_subnet1             ... bench:         106 ns/iter (+/- 7)
test parse_one_subnet2             ... bench:          94 ns/iter (+/- 0)
test test_10000_ips_in_chnlists    ... bench:     429,550 ns/iter (+/- 4,530)
```

It it about 3x faster to add subnets and more than 10x faster to test whether an ip is in.